refactor to force dispatching on out datatype
all ops now create their output holder and pass that to the type
dispatcher. Some (most) ops don't have particularly intelligent
output holder creation, but the point of this is to allow providing
an existing output array or forcing the output array to a specific
type (for example, boolean comparison ops)

This was in prep for boolean comparison ops so we can compare
to thresholds in bliss

This does seem to make things compile way slower... binary ops
seem to cause the most pain, especially when ndarray = f(ndarray,
ndarray). OTOH, I get it because that's a 3-way dispatch and all code
paths for all types needs to be added, but sheesh we're now at 4 min
and trying to slash out code that wasn't totally needed doesn't seem to
make a huge difference.