file reading bug fixes:
fixes a bug with creating large tensors where the number of elements
could overflow the (accidental) default int value. Change to 1LL
to get the correct product in this case

catch a bug with some data files which has the DIMENSION_LABELS swapped
for channels and time. Check the reported nchans and if it matches
the reported time dimension, swap the labels