Desmear noise comp (#9)
pipeline: compensating SNR for desmeared integrations

Add a class to capture all of the metadata outputs along with the actual dedrift integration
from the integration phase. This includes how much each rate has been desmeared. That's used
in the hit-finding processes to adjust noise estimates based on the total number of bins used
in an integration by dividing by sqrt(.) of that number. This is an improvement over the previous
state of using just the number of spectra to adjust nosie powers.