setup easier API for scanning multiple coarse channels
Moves the structure (map) holding coarse channels to a shared_ptr<map>
so slices of a scan share the same holder. This means when requesting
hits from all coarse_channels from the root scan it will have access
to the coarse_channels that have gone through the pipeline.

Also add a peak method for coarse_channels that will only return a valid
coarse_channel if it's already in the holder and rename the `get_coarse_channel`
method to a now more clear `read_coarse_channel` since that will trigger
a read from disk if not already in the map

Also adds minor plot utils to python package