fix bliss_find_hits cli and begin alternate integration kernel
the integration kernel appears slower than it could be by probably 100msec
or so. Was comparing to hyperseti, and decided the major difference is the
scrunching or reduction of number of drifts searched rather than some intrinsic
property of the kernel. The new one can be faster, but doesn't yet support
full range of features, so it's left in but isn't used