telescopes: fix metadata mixups and channelizations
ATA uses a bytestring rather than an encoded string for the source_name
in hdf5. This adds a check if a string md field is actually a bytestring
and converts it to a string as needed.

Add support for slicing to the end and slicing channels in the bliss_find_hits
executable

Removes the non-channelization specific ctors for cadence, observation_target, scan
and replaces them with the 0-defaulted channelization-specific versions with
friendlier pybinds