create mask when not present in hdf5 (#67)
* create mask when not present in hdf5

Closes #65. Some files (many of the recent files) do not have a `mask`
dataset in the hdf5 file. It's always uint8 and zero, so this sets the
hdf5 mask handle as optional and just creates a zero-valued mask of
appropriate size when it's not present.

Also  add support for hdf5 filterbank files with VERSION == 2.0 which was added
through blimpy to add the nfpc field which gives the fine channels per coarse.
N.B.: this isn't yet **using** the nfpc field but accepts VERSION being 2.0 as not
an error and will generally issue a WARN instead of throw an exception when the
VERSION field does not match a known version.