fix spectral kurtosis d (shape parameter to gamma distribution) default value to be 2
This exposes the `d` parameter from the generalized spectral kurtosis estimator and
sets the default value to 2. Previously it was parameterized internally but set to 1
to match another implementation. I believe the default value for dual-polarization spectrum
should be 2 to give a gamma distribution with shape that matches chi-square with 4 degrees
of freedom.